vincula candidato e usuario

+++ src/main/java/br/com/cristal/erp/service/candidato/CandidatoService.java
@@ -1,13 +1,17 @@
 package br.com.cristal.erp.service.candidato;
 
+import br.com.cristal.erp.config.usuarioConfig.CustomUserDetailsService;
 import br.com.cristal.erp.controller.candidato.dto.CandidatoPostRequestBody;
 import br.com.cristal.erp.controller.candidato.dto.CandidatoPutRequestBody;
 import br.com.cristal.erp.controller.candidato.dto.CandidatoResponseBody;
+import br.com.cristal.erp.exception.AcessDeniedException;
 import br.com.cristal.erp.exception.BadRequestsException;
 import br.com.cristal.erp.mapper.CandidatoMapper;
 import br.com.cristal.erp.repository.candidato.CandidatoRepository;
 import br.com.cristal.erp.repository.candidato.model.Candidato;
 import br.com.cristal.erp.repository.candidato.model.enums.StatusCandidato;
+import br.com.cristal.erp.repository.usuario.model.Usuario;
+import br.com.cristal.erp.util.JWTUtility;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -20,14 +24,27 @@ public class CandidatoService {
 
     private CandidatoRepository candidatoRepository;
     private CandidatoMapper candidatoMapper;
+    private JWTUtility jwtUtility;
+    private CustomUserDetailsService customUserDetailsService;
 
-    public CandidatoResponseBody replace(CandidatoPutRequestBody candidatoToBeUpdated) {
 
-        Candidato candidatoFound = findByIdOrThrowBadRequestException(candidatoToBeUpdated.getId());
-        Candidato candidatoToBeSaved = CandidatoMapper.INSTANCE.toCandidato(candidatoToBeUpdated);
-        candidatoToBeSaved.setId(candidatoFound.getId());
-        Candidato savedCandidato = candidatoRepository.save(candidatoToBeSaved);
-        return CandidatoMapper.INSTANCE.toResponseBody(savedCandidato);
+    public CandidatoResponseBody replace(String headerToken, CandidatoPutRequestBody requestPutCandidato) {
+
+        // pega usuário pelo email
+        Usuario usuario = customUserDetailsService
+                .loadUserByEmailAndReturnsUsuario(
+                        jwtUtility.getEmailFromToken(headerToken.substring(7))
+                );
+
+        Candidato candidatoFound = findByIdOrThrowBadRequestException(usuario.getId());
+        Candidato candidatoToBeUpdated = CandidatoMapper.INSTANCE.toCandidato(requestPutCandidato);
+
+        candidatoToBeUpdated.setId(candidatoFound.getId());
+
+        Candidato updatedCandidato = candidatoRepository.save(candidatoToBeUpdated);
+
+        return CandidatoMapper.INSTANCE.toResponseBody(updatedCandidato);
+
     }
 
     public Candidato findByIdOrThrowBadRequestException(long id){
@@ -40,13 +57,35 @@ public class CandidatoService {
         return CandidatoMapper.INSTANCE.toResponseBody(candidato);
     }
 
-    public CandidatoResponseBody save(CandidatoPostRequestBody candidatoPost){
-        Candidato candidato = CandidatoMapper.INSTANCE.toCandidato(candidatoPost);
-        candidato = candidatoRepository.save(candidato);
+    public CandidatoResponseBody save(String headerToken, CandidatoPostRequestBody candidatoPost){
+
+        // pega usuário pelo email
+        Usuario usuario = customUserDetailsService
+                .loadUserByEmailAndReturnsUsuario(
+                    jwtUtility.getEmailFromToken(headerToken.substring(7))
+                );
+
+        // excessão caso já esteja cadastrado
+        if(candidatoRepository.findById(usuario.getId()).isPresent()){
+            throw new BadRequestsException("Candidato Já Cadastrado");
+        }
+
+        Candidato candidatoToBeSaved = CandidatoMapper.INSTANCE.toCandidato(candidatoPost);
+        candidatoToBeSaved.setId(usuario.getId());
+
+        Candidato candidato = candidatoRepository.save(candidatoToBeSaved);
+
         return CandidatoMapper.INSTANCE.toResponseBody(candidato);
     }
 
-    public void delete(Long id){
+    public void delete(String headerToken, Long id){
+        String userEmail = jwtUtility.getEmailFromToken(headerToken.substring(7));
+        Usuario usuario = customUserDetailsService.loadUserByEmailAndReturnsUsuario(userEmail);
+
+        if(!id.equals(usuario.getId())){
+            throw new AcessDeniedException("Permissão Negada");
+        }
+
         Candidato candidatoToBeDeleted = findByIdOrThrowBadRequestException(id);
         candidatoRepository.delete(candidatoToBeDeleted);
     }
